+++ docs/site-config.js
@@ -1,8 +1,8 @@
 window.MINOVA_CINEMA_CONFIG = Object.freeze({
-  currentVersion: "2.5.0",
+  currentVersion: "2.5.1",
   feedbackEndpoint: "",
   feedbackRecipient: "minova.chromium@gmail.com",
   repositoryUrl: "https://github.com/minova-chromium/Minova-Android-Tv-Cinema-Application",
   latestReleaseApiUrl: "https://api.github.com/repos/minova-chromium/Minova-Android-Tv-Cinema-Application/releases/latest",
-  latestApkUrl: "https://github.com/minova-chromium/Minova-Android-Tv-Cinema-Application/releases/download/v2.5.0/Minova-Cinema-2.5.0.apk"
+  latestApkUrl: "https://github.com/minova-chromium/Minova-Android-Tv-Cinema-Application/releases/download/v2.5.1/Minova-Cinema-2.5.1.apk"
 });